Sort files by name Delete files and after create again

+++ tools/pipelines-maven-plugin/src/main/java/org/gbif/pipelines/maven/AvroPostprocessMojo.java
@@ -111,11 +111,19 @@ public class AvroPostprocessMojo extends AbstractMojo {
   private void writeFile(Path path, List<String> lines, List<Integer> idxs) {
     if (idxs.get(0) != -1 || idxs.get(1) != -1) {
       try {
-        Files.write(path, lines);
+        if (Files.deleteIfExists(path)) {
+          Files.write(path, lines);
+          if (Files.exists(path)) {
+            getLog().info("Modified - " + path.toString());
+          } else {
+            getLog().error("Can't create file - " + path.toString());
+          }
+        } else {
+          getLog().error("Can't modify - " + path.toString());
+        }
       } catch (IOException ex) {
         throw new IllegalStateException(ex.getMessage(), ex);
       }
-      getLog().info("Modified - " + path.toString());
     }
   }
 
@@ -228,7 +236,7 @@ public class AvroPostprocessMojo extends AbstractMojo {
   /** Searches for all java classes in a directory */
   private List<Path> searchClasses() {
     try (Stream<Path> paths = Files.walk(Paths.get(directory))) {
-      return paths.filter(path -> path.toFile().isFile()).collect(Collectors.toList());
+      return paths.filter(path -> path.toFile().isFile()).sorted().collect(Collectors.toList());
     } catch (IOException ex) {
       throw new IllegalStateException(ex.getMessage(), ex);
     }
